Fix: Add Lottie, Color Extention
