+++ Harmonica/Harmonica/View/Main/MainView.swift
@@ -58,22 +58,6 @@ struct MainView: View {
     }
 }
 
-extension Color {
-    init(hex: String) {
-        let scanner = Scanner(string: hex)
-        _ = scanner.scanString("#") // "#" 기호 제거
-        
-        var rgb: UInt64 = 0
-        scanner.scanHexInt64(&rgb)
-        
-        let r = Double((rgb >> 16) & 0xFF) / 255.0
-        let g = Double((rgb >> 8) & 0xFF) / 255.0
-        let b = Double(rgb & 0xFF) / 255.0
-        
-        self.init(red: r, green: g, blue: b)
-    }
-}
-
 struct SongBox:View {
     @State var SongName:String = ""
     @State var AlbumCover:Image?
