+++ Harmonica/Harmonica/src/Media.xcassets/Button/STTSearchView_Unpressed.imageset/Contents.json
